refactor: Notification 컴포넌트 구조 개선
- Notification 범용 UI 컴포넌트 분리
- NotificationList를 app/components로 이관
