+++ src/basic/app/components/NotificationList.tsx
@@ -1,0 +1,20 @@
+import { Notification, type NotificationItem } from "../../shared";
+
+interface NotificationListProps {
+  notifications: NotificationItem[];
+  onRemove: (id: string) => void;
+}
+
+export function NotificationList({ notifications, onRemove }: NotificationListProps) {
+  if (notifications.length === 0) {
+    return null;
+  }
+
+  return (
+    <div className="fixed right-4 top-20 z-50 max-w-sm space-y-2">
+      {notifications.map((notif) => (
+        <Notification {...notif} onRemove={onRemove} />
+      ))}
+    </div>
+  );
+}
