+++ src/basic/app/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./Header";
+export * from "./NotificationList";
