+++ src/basic/shared/components/ui/index.ts
@@ -1,4 +1,4 @@
 export * from "./BadgeContainer";
 export * from "./Button";
-export * from "./NotificationList";
+export * from "./Notification";
 export * from "./SearchInput";
