+++ src/basic/shared/types/notification.ts
@@ -1,4 +1,4 @@
-export interface Notification {
+export interface NotificationItem {
   id: string;
   message: string;
   type: "error" | "success" | "warning";
